refactor: AssetVault のリトライループを AssetVaultRetry に共通化
