+++ Assets/UniLab/AssetVault/AssetVaultRuntime.cs
@@ -8,11 +8,11 @@ namespace UniLab.AssetVault
         /// <summary>
         /// 環境ごとのホスト込み基底 URL です（例 https://dev1.xxx.xxx/app）。env から URL へのマッピングはアプリ config が持ちます。
         /// </summary>
-        public static string BaseUrl { get; set; } = null;
+        public static string BaseUrl { get; set; }
 
         /// <summary>
         /// version.json の path です。RemoteLoadPath の不透明な版セグメントに使われます。
         /// </summary>
-        public static string ContentPath { get; set; } = null;
+        public static string ContentPath { get; set; }
     }
 }
